+++ src/Service/Controller/Other/AdminAccount.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Service\Controller;
+namespace App\Service\Controller\Other;
 
 use App\Model\Table\Shared\AccountStatusMastersTable;
 use App\Service\Controller\Shared\ServiceInterface;
